# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for Return Zero music app.",
        "title": "Return Zero API",
        "contact": {},
        "version": "1.0"
    },
    "host": "returnzero.ru",
    "basePath": "/",
    "paths": {
...
        "/artists/{id}/tracks": {
            "get": {
                "description": "Get a list of tracks by a specific artist with optional pagination filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "Get tracks by artist ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Artist ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tracks by artist",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/delivery.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/delivery.Track"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid ID or filters",
                        "schema": {
                            "$ref": "#/definitions/delivery.APIBadRequestErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/delivery.APIInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/tracks": {
            "get": {
                "description": "Get a list of tracks with optional pagination filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "Get tracks",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Offset (default: 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit (default: 10, max: 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tracks",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/delivery.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/delivery.Track"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid filters",
                        "schema": {
                            "$ref": "#/definitions/delivery.APIBadRequestErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/delivery.APIInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/tracks/{id}": {
            "get": {
                "description": "Retrieves a specific track by its ID with detailed information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tracks"
                ],
                "summary": "Get track by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Track ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Track details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/delivery.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/delivery.TrackDetailed"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid ID",
                        "schema": {
                            "$ref": "#/definitions/delivery.APIBadRequestErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/delivery.APIInternalServerErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "delivery.APIBadRequestErrorResponse": {
            "description": "API bad request error response structure",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Something went wrong"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
        "delivery.APIInternalServerErrorResponse": {
            "description": "API internal server error response structure",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Something went wrong"
                },
                "status": {
                    "type": "integer",
                    "example": 500
                }
            }
        },
        "delivery.APIResponse": {
            "description": "API response wrapper",
            "type": "object",
            "properties": {
                "body": {},
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
        "delivery.Track": {
            "description": "A music track entity",
            "type": "object",
            "properties": {
                "album": {
                    "type": "string"
                },
                "album_id": {
                    "type": "integer",
                    "example": 1
                },
                "artists": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/delivery.TrackArtist"
                    }
                },
                "duration": {
                    "type": "integer",
                    "example": 216
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "thumbnail_url": {
                    "type": "string",
                    "example": "https://example.com/image.jpg"
                },
                "title": {
                    "type": "string",
                    "example": "Lagtrain"
                }
            }
        },
        "delivery.TrackArtist": {
            "description": "An artist associated with a track",
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "role": {
                    "type": "string",
                    "example": "Main artist"
                },
                "title": {
                    "type": "string",
                    "example": "Lagtrain"
                }
            }
        },
        "delivery.TrackDetailed": {
            "type": "object",
            "properties": {
                "album": {
                    "type": "string"
                },
                "album_id": {
                    "type": "integer",
                    "example": 1
                },
                "artists": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/delivery.TrackArtist"
                    }
                },
                "duration": {
                    "type": "integer",
                    "example": 216
                },
                "file_url": {
                    "type": "string",
                    "example": "https://example.com/track.mp3"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "thumbnail_url": {
                    "type": "string",
                    "example": "https://example.com/image.jpg"
                },
                "title": {
                    "type": "string",
                    "example": "Lagtrain"
                }
            }
        }
    }
}